Update tilix to 1.9.4
Summary:
The old autotools build system has been removed in favor of Meson/Ninja. The previous patch file has been dropped since it is in the release. The new patch is to strip the step using `appdatacli`, since we don't appear to have it in Solus at the moment.

**Summarized Changelog:**
- **Added:**
  - Option to strip trailing whitespace on paste
  - Shortcut to "Unselect all"

- **Fixed:**
  - Actually install Yaru color scheme
  - Saving of already saved session

- **Changed:**
  - Give every tab the `${title}` instead of "Default"

Test Plan: Change color profiles, change directories, split horizontally and vertically, and make and send this patch

Reviewers: #triage_team, JoshStrobl

Reviewed By: #triage_team, JoshStrobl

Subscribers: JoshStrobl

Differential Revision: https://dev.getsol.us/D10521